Natural_disaster/requirements.txt
torch
torchvision
numpy
scikit-learn
matplotlib
tqdm
Pillow
opencv-python
reportlab